requirements.txt
# Flask PWA stack
flask==3.0.0
gunicorn==21.2.0

# Google Cloud APIs (mantido para OCR)
google-cloud-vision==3.4.5
google-auth==2.29.0
google-generativeai==0.3.2

# Environment & Security
python-dotenv==1.0.0

# Image Processing (compatível com Python 3.13)
Pillow==10.4.0

# HTTP client
httpx==0.27.0

# Telegram Bot (Multi-Entregador)
python-telegram-bot==20.7

# Parsers de Romaneios
pdfplumber==0.11.0  # PDF com texto extraível
PyPDF2==3.0.1  # Alternativa para PDF
pytesseract==0.3.10  # OCR para PDFs escaneados (requer Tesseract instalado)
pdf2image==1.17.0  # Converte PDF em imagem para OCR